% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{generate_simulation_data}
\alias{generate_simulation_data}
\title{Generate Simulated Data for Time-to-Event and Longitudinal Analyses}
\usage{
generate_simulation_data(
  fmla.tte,
  fmla.long,
  bootstrap = FALSE,
  bootstrapfrom = NULL,
  id.name = "id",
  beta.tte,
  scale.tte = NULL,
  shape.tte = NULL,
  normal.tte = FALSE,
  sd.tte = NULL,
  beta.y,
  sd.y,
  randeff.mean = c(0.5, 0, -1, 1),
  randeff.sd = rep(0.2, 4),
  randeff.corr = NULL,
  n = 100,
  censor.parameter,
  time.interval = 0.1,
  time.interval.sd = 0.02,
  seed = 1
)
}
\arguments{
\item{fmla.tte}{A formula object that delineates the structure of the time-to-event model.}

\item{fmla.long}{Another formula object that outlines the structure of the longitudinal data model.}

\item{bootstrap}{A logical value indicating whether to bootstrap the data. If TRUE, the function will sample from the dataset specified in bootstrapfrom. Default is FALSE.}

\item{bootstrapfrom}{A data frame. When bootstrap is TRUE, this data frame is used for resampling. Default is NULL.}

\item{id.name}{A character string specifying the name of the ID column in the bootstrapfrom dataset. Default is "id".}

\item{beta.tte}{A numeric vector that defines the beta coefficients for the TTE model.}

\item{scale.tte}{A single numeric value that sets the scale parameter for the TTE model.}

\item{shape.tte}{A single numeric value that sets the shape parameter for the TTE model.}

\item{normal.tte}{Logical indicating if the time-to-event model should follow a log normal distribution. Default is FALSE.}

\item{sd.tte}{Optional numeric value representing the standard deviation for the time-to-event model. Only used if normal.tte is TRUE.}

\item{beta.y}{A numeric vector that defines the beta coefficients for the longitudinal data model.}

\item{sd.y}{A single numeric value, representing the standard deviation of the Y variable in the longitudinal data model.}

\item{randeff.mean}{A numeric vector that defines the average of the random effects.}

\item{randeff.sd}{A numeric vector that designates the standard deviations associated with the random effects.}

\item{randeff.corr}{A matrix that sets the correlation structure of the random effects. If NULL, a random structure will be generated.}

\item{n}{An integer that represents the number of observations to simulate.}

\item{censor.parameter}{A numeric value used to designate the parameter for the exponential distribution, which is employed to derive censoring times.}

\item{time.interval}{A numeric value that stipulates the expected time interval for visits in the longitudinal data.}

\item{time.interval.sd}{A numeric value representing the standard deviation for the time interval of visits in the longitudinal data. Default is 0.02.}

\item{seed}{An integer used to set the seed for random number generation, ensuring reproducibility.}
}
\description{
The generate_simulation_data function provides a convenient way to produce simulated data
tailored for both time-to-event and longitudinal data analyses. This can be particularly useful for methodological studies,
preliminary analyses, and educational purposes.
}
\section{Parameters}{

}

\section{Returns}{

The function returns a list comprising three elements:
\itemize{
\item \code{survdat}: A data frame containing the simulated survival data.
\item \code{longdat}: A data frame encompassing the simulated longitudinal data.
\item \code{simulation.para}: A list that consolidates the parameters applied during the simulation.
}
}

\section{Examples}{

\preformatted{
simulation.list <- generate_simulation_data(
fmla.tte = as.formula(Surv(PFS_YEARS, PFS_EVENT) ~ Y0SCALE + Y1SCALE),
fmla.long = as.formula(PCHG ~ 0 + Y0SCALE + Y2SCALE + Y3),

beta.tte = c(0.1, 0.05, 0.1), scale.tte = 2, shape.tte = 2,
beta.y = c(0.02, -0.02, 0.03), sd.y = 0.1,
randeff.mean = c(0.5, 0, -1, 1), randeff.sd = rep(0.2, 4), randeff.corr = NULL,
n = 1000, censor.parameter = 2, normal.tte = FALSE, time.interval = 0.1
)
}
}

