% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataloader.R
\name{load_onearm_data}
\alias{load_onearm_data}
\title{Load and Process Change Point Data for One Arm}
\usage{
load_onearm_data(
  survdat,
  longdat,
  fmla.tte,
  fmla.long,
  longdat.time,
  id.indicator = "id",
  chgpt_location = 0.5,
  b0_location = 0,
  b1_location = -0.5,
  b2_location = 0.5,
  chgpt_scale = 0.5,
  b0_scale = 1,
  b1_scale = 0.5,
  b2_scale = 0.5,
  chgpt_shape = 8,
  b0_shape = 8,
  b1_shape = 8,
  b2_shape = 8
)
}
\arguments{
\item{survdat}{A data frame containing the pseudo survival data.}

\item{longdat}{A data frame containing the pseudo longitudinal data.}

\item{fmla.tte}{A formula defining the structure for time-to-event analysis.}

\item{fmla.long}{A formula defining the structure for longitudinal analysis.}

\item{longdat.time}{A character string specifying the time variable in the longitudinal data.}

\item{id.indicator}{A character string indicating the column name for the ID indicator in both `survdat` and `longdat`. Default is "id".}

\item{chgpt_location, }{b0_location, b1_location, b2_location Initial values for the location parameters. Default values are 0.5, 0, -0.5, and 0.5, respectively.}

\item{chgpt_scale, }{b0_scale, b1_scale, b2_scale Initial values for the scale parameters. Default values are 0.5, 1, 0.5, and 0.5, respectively.}

\item{chgpt_shape, }{b0_shape, b1_shape, b2_shape Initial values for the shape parameters. Default values are all 8.}
}
\value{
A list containing processed data ready for Bayesian change point analysis, including segmented observed and censored data,
new id indicators, and derived model matrices based on the input formulas for time-to-event and longitudinal analyses.
}
\description{
This function takes in pseudo survival and longitudinal data, processes them for a specific arm,
and prepares it for Bayesian change point analysis. The data processing includes segregating
observed and censored data, generating new id indicators, and summarizing data for the respective arm.
}
\examples{
# Load required datasets
data("survdat")
data("longdat")

# Define a formula for time-to-event analysis
example_formula_tte <- as.formula(Surv(PFS_YEARS, PFS_EVENT) ~ 0 + Y0SCALE)

# Define a formula for longitudinal analysis
example_formula_long <- as.formula(PCHG ~ 0 + Y0SCALE)

# Use the load_onearm_data function
standat.list <- load_onearm_data(
  survdat = survdat,
  longdat = longdat,
  fmla.tte = example_formula_tte,
  fmla.long = example_formula_long,
  longdat.time = "YEAR",
  id.indicator = "id"
)

}
