% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{generate_simulation_data}
\alias{generate_simulation_data}
\title{Generate Simulation Data for Time-to-Event and Longitudinal Data Analyses}
\usage{
generate_simulation_data(
  fmla.tte,
  fmla.long,
  betatte,
  scalette = NULL,
  shapette = NULL,
  normal.tte = FALSE,
  sd_tte = NULL,
  beta_y,
  sd_y,
  randeff.mean = c(0.5, 0, -1, 1),
  randeff.sd = rep(0.2, 4),
  randeff.corr = NULL,
  n = 100,
  censor.parameter,
  time.interval = 0.1,
  seed = 1
)
}
\arguments{
\item{fmla.tte}{A formula object specifying the structure of the time-to-event model.}

\item{fmla.long}{A formula object specifying the structure of the longitudinal data model.}

\item{betatte}{A numeric vector specifying the beta parameters for the TTE model.}

\item{scalette}{A numeric value specifying the scale parameter for the TTE model.}

\item{shapette}{A numeric value specifying the shape parameter for the TTE model.}

\item{beta_y}{A numeric vector specifying the beta parameters for the longitudinal data model.}

\item{sd_y}{A numeric value specifying the standard deviation of the Y variable in the longitudinal data model.}

\item{randeff.mean}{A numeric vector specifying the mean of the random effects.}

\item{randeff.sd}{A numeric vector specifying the standard deviations of the random effects.}

\item{randeff.corr}{A matrix specifying the correlation structure of the random effects, or NULL to generate a random structure.}

\item{n}{An integer specifying the number of observations to simulate.}

\item{censor.parameter}{A numeric value specifying the parameter for the exponential distribution used to generate censoring times.}

\item{time.interval}{A numeric value specifying the time interval for visits in the longitudinal data.}

\item{seed}{An integer specifying the seed for random number generation.}
}
\value{
A list containing three elements:
\itemize{
 \item \code{survdat} A data frame containing the simulated survival data.
 \item \code{longdat} A data frame containing the simulated longitudinal data.
 \item \code{simulation.para} A list containing the parameters used in the simulation.
}
}
\description{
This function generates simulated data suitable for time-to-event and longitudinal data analyses,
based on the provided parameters and formulae.
}
\examples{
simulation.list <- generate_simulation_data(
  fmla.tte = as.formula(Surv(PFS_YEARS, PFS_EVENT) ~ Y0SCALE + Y1SCALE),
  fmla.long = as.formula(PCHG ~ 0 + Y0SCALE + Y2SCALE + Y3),
  betatte = c(0.1, 0.05, 0.1), scalette = 2, shapette = 2,
  beta_y = c(0.02, -0.02, 0.03), sd_y = 0.1,
  randeff.mean = c(0.5, 0, -1, 1), randeff.sd = rep(0.2, 4), randeff.corr = NULL,
  n = 1000, censor.parameter = 2, time.interval = 0.1
)
}
